pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-traces"
version = "0.1.0"
description = "Universal LLM and agent tracing for Python agent frameworks"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["agents", "observability", "langchain", "tracing", "llm"]
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.46.0",
]

[project.optional-dependencies]
langchain = [
    "deepagents>=0.5.8",
    "langchain>=1.2.17",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "python-dotenv>=1.2.2",
]
crewai = [
    "crewai>=1.8.0",
    "litellm>=1.80.0",
]
autogen = [
    "pyautogen>=0.10.0",
]
openai-agents = [
    "openai-agents>=0.6.0",
]
postgres = [
    "psycopg[binary]>=3.2.0",
]
all = [
    "langchain>=1.2.17",
    "langchain-openai>=1.2.1",
    "python-dotenv>=1.2.2",
    "crewai>=1.8.0",
    "litellm>=1.80.0",
    "pyautogen>=0.10.0",
    "openai-agents>=0.6.0",
    "psycopg[binary]>=3.2.0",
]

[project.scripts]
agent-obs = "universal_agent_obs.server.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["universal_agent_obs*", "samples*"]

[tool.setuptools.package-data]
"universal_agent_obs.server.static" = ["*.html"]
